Adding script to extract accessions & count reads
This is a script that Chris Grahlmann wrote to extract accession numbers from the UnambiguouslyMapped* fastq files and count the number of reads that were simulated from each accession. It is useful for obtaining accession numbers and verifying the abundance of each organism within the simulated sample. It is also useful for generating ground truth information for UnAmbiguouslyMapped_ds.frankengenome.fq.gz, which was not included in the truth set information for the McIntyre study.

We ran the script with these commands (note that the script requires *.fq and does not accept *.fq.gz): 
python strip_fq_files.py --file UnAmbiguouslyMapped_ds.7.fq > ds7_counts.txt

This script will output the accession number followed by the number of reads that were simulated from it:
NC_016111.1 4149
NC_012917.1 5364
NC_016751.1 3434
NC_016113.1 4859
NZ_ACZP01000022.1 5179
NC_014005.1 310
NZ_ADKR01000377.1 451
NZ_ACPX01000184.1 1160
NC_011990.1 1323
NC_014007.1 2785

Rachid Ounit, the original author of these datasets, recommended that we use the accession numbers in the fastq files to update the NCBI taxonomy for each organism. When this was done in March 2018 and compared to the truth information published in the McIntyre et al. 2017 study (https://ftp-private.ncbi.nlm.nih.gov/nist-immsa/IMMSA/truth_sets.zip), the following discrepancies were identified:

1) Within the NYCSM dataset, the GenBank accession numbers NZ_AUTP01000201.1 - NZ_AUTP01000216.1 correspond to the species Pseudoalteromonas sp. TAB23 (NCBI Taxonomy ID = 1938595). This is different than the species and strain listed in the truth set for McIntyre et al. 2017 - Pseudoalteromonas haloplanktis (NCBI Taxonomy ID = 228) and Pseudoalteromonas haloplanktis TAB23 (NCBI Taxonomy ID = 1333511).

2) Within the ds.7 dataset, the GenBank accession number NC_009379.1 corresponds to the species Polynucleobacter asymbioticus (NCBI Taxonomy ID = 576611) and no rank Polynucleobacter asymbioticus QLW-P1DMWA-1 (NCBI Taxonomy ID = 312153). This is different than the species listed in the truth set for McIntyre et al. 2017 - Polynucleobacter necessarius (NCBI Taxonomy ID = 576610), but the subspecies listed in the truth set for McIntyre et al. 2017 has the same NCBI Taxonomy ID with a different name - Polynucleobacter necessarius subsp. asymbioticus (NCBI Taxonomy ID = 576611).

Note that this script does not identify reads that no longer unambiguously map to a single species. This has happened since the publication of the McIntyre study in 2017, since new near neighbors of the organisms within these simulated datasets continue to be added to RefSeq.